+++ src/main/UIGFJson.js
@@ -6,12 +6,15 @@ const getItemTypeNameMap = require('./gachaTypeMap').getItemTypeNameMap
 const { version } = require('../../package.json')
 const config = require('./config')
 const fetch = require('electron-fetch').default
-const { readJSON, saveJSON, existsFile, userDataPath, fixLocalMap } = require('./utils')
+const { readJSON, saveJSON, existsFile, userDataPath, fixLocalMap, md5, mapToObject } = require('./utils')
 const Ajv = require('ajv')
+const Ajv2020 = require('ajv/dist/2020')
 
 // acquire uigf schema
-const validateUigfJson = new Ajv().compile(require('../schema/uigf.json'))
-const validateLocalJson = new Ajv().compile(require('../schema/local-data.json'))
+const validateUigf30Json = new Ajv({ strict: false }).compile(require('../schema/uigf3_0.json'))
+const validateUigf41Json = new Ajv2020({ strict: false }).compile(require('../schema/uigf4_1.json'))
+const validateUigf42Json = new Ajv2020({ strict: false }).compile(require('../schema/uigf4_2.json'))
+const validateLocalJson = new Ajv({ strict: false }).compile(require('../schema/local-data.json'))
 
 const uigfLangMap = new Map([
   ['zh-cn', 'chs'],
@@ -21,13 +24,14 @@ const uigfLangMap = new Map([
   ['es-es', 'es'],
   ['fr-fr', 'fr'],
   ['id-id', 'id'],
-  ['ja-jp', 'ja'],
+  ['ja-jp', 'jp'],
   ['ko-kr', 'kr'],
   ['pt-pt', 'pt'],
   ['ru-ru', 'ru'],
   ['th-th', 'th'],
   ['vi-vn', 'vi']
 ])
+const uigfRevLangMap = new Map(Array.from(uigfLangMap, ([key, value]) => [value, key]))
 
 const getTimeString = () => {
   return new Date().toLocaleString('sv').replace(/[- :]/g, '').slice(0, -2)
@@ -58,70 +62,86 @@ const shouldBeString = (value) => {
 // a dictionary for looking up item ids
 const itemIdDict = new Map()
 // the md5 value for the dictionary
-let itemIdDictMd5 = null
+const itemIdDictMd5 = new Map()
 // the file name for saving item id dictionary
 const itemIdDictFileName = 'item-id-dict.json'
 
 // acquire dictionary based on give language
-const fetchItemIdDict = async (lang = 'all') => {
-  // fetch item id dict from api.uigf.org
-  const response = await fetch(`https://api.uigf.org/dict/genshin/${lang}.json`)
-  // update local dict
-  Object.entries(await response.json()).forEach(
-    ([lang, table]) => itemIdDict.set(
-      // assign item id table based on language, and convert all id from number to string
-      lang, new Map(Object.entries(table).map(([name, id]) => [name, String(id)]))
-    )
-  )
+const fetchItemIdDict = async () => {
+  // fetch item id dicts from api.uigf.org
+  let response = await fetch(`https://api.uigf.org/dict/genshin/all.json`)
+  let responseJson = await response.json()
+  if (response.status === 200) { // successfully fetched all.json
+    Object.entries(responseJson).forEach(([lang, dict]) => {
+      itemIdDict.set(lang, new Map(Object.entries(dict).map(([name, id]) => [name, String(id)])))
+    })
+  } else { // all.json failed, attempting to fetch individual languages
+    console.error(`All dict fetch failed: HTTP ${response.status}: ` + (responseJson.detail || 'unknown error'))
+    for (const [_, lang] of uigfLangMap) {
+      response = await fetch(`https://api.uigf.org/dict/genshin/${lang}.json`)
+      const responseText = await response.text().catch((e) => console.error(e))
+      responseJson = JSON.parse(responseText)
+      if (response.status !== 200) {
+        console.error(`${lang} dict fetch failed: HTTP ${response.status}: ` + (responseJson.detail || 'unknown error'))
+        continue
+      }
+      itemIdDict.set(lang, new Map(Object.entries(responseJson).map(([name, id]) => [name, String(id)])))
+      itemIdDictMd5.set(lang, md5(responseText))
+    }
+  }
 }
 
 // acquire dictionary based on give language
 const fetchItemIdDictMd5 = async (lang = 'all') => {
   const response = await fetch('https://api.uigf.org/md5/genshin')
   const responseJson = await response.json()
-  return responseJson[lang]
+  if (response.status !== 200) throw new Error(`MD5 fetch failed: HTTP ${response.status}: ` + (responseJson.detail || 'unknown error'))
+  Object.entries(responseJson).forEach(([lang, hash]) => itemIdDictMd5.set(lang, hash))
 }
 
 // initialize item id dictionary
 const initLookupTable = async () => {
-  // if itemIdDictMd5 != null, that means itemIdDict has been init, no need to do it again
-  if (itemIdDictMd5) {
+  // check if itemIdDict has been init, no need to do it again
+  if (itemIdDict.size > 0) {
     return
   }
 
-  // try to obtain dict md5
-  try {
-    itemIdDictMd5 = await fetchItemIdDictMd5()
-  } catch (e) {
-    console.log(`Unable to fetch latest item id dictionary md5 due to: ${e}`)
-  }
+  // fetch the remote MD5 hashes
+  await fetchItemIdDictMd5().catch((e) => console.error(e))
 
-  // if a locally cached dictionary does not exist
+  // if a locally cached dictionary does not exist, fetch a new copy
   if (!existsFile(itemIdDictFileName)) {
     await fetchItemIdDict();
     return;
   }
 
-  // if a locally cached dictionary is found
-  const data = await readJSON(itemIdDictFileName)
-  // if itemIdDictMd5 is not successfully fetched previously
-  if (!itemIdDictMd5 && data) itemIdDictMd5 = data.md5
+  // load the locally cached dictionary
+  const localItemIdDict = await readJSON(itemIdDictFileName)
 
-  // if the data is null or the md5 does not match
-  if (!data || data.md5 !== itemIdDictMd5) {
-    // console.log('md5 check failed! Re-fetching...')
+  // if localItemIdDict is empty or old version, refetch
+  if (!localItemIdDict || typeof localItemIdDict.md5 === "string") {
     await fetchItemIdDict()
-    return;
+    return
   }
 
-  // if the data is valid and the md5 matches
-  // console.log('md5 check success!')
-  data.lang.forEach(([lang, table]) => itemIdDict.set(lang, new Map(table)))
+  // ensure all remote md5 matches local hashes
+  for (const [lang, hash] of itemIdDictMd5.entries()) {
+    const localHash = localItemIdDict.md5[lang] || ''
+    if (localHash !== hash) {
+      await fetchItemIdDict()
+      return
+    }
+  }
+
+  // insert locally cached dicts into memory
+  for (const [lang, dict] of Object.entries(localItemIdDict.lang)) {
+    itemIdDict.set(lang, new Map(Object.entries(dict).map(([name, id]) => [name, String(id)])))
+  }
 }
 
 // save item id dictionary
 const saveLookupTable = async () => {
-  await saveJSON(itemIdDictFileName, { lang: itemIdDict, md5: itemIdDictMd5 })
+  await saveJSON(itemIdDictFileName, { lang: mapToObject(itemIdDict), md5: mapToObject(itemIdDictMd5) })
 }
 
 // get item id
@@ -130,10 +150,10 @@ const getItemId = async (lang, name) => {
   if (!itemIdDict.has(lang) || !itemIdDict.get(lang).has(name)) {
     const response = await fetch(`https://api.uigf.org/identify/genshin/${name}`)
     const responseJson = await response.json()
-    if (!responseJson.item_id) {
+    if (response.status != 200) {
       throw new Error(`Couldn't find the item_id for the ${name}.`)
     }
-    itemIdDict.get(lang).set(name, responseJson.item_id.toString())
+    itemIdDict.get(lang).set(name, responseJson.matched[0].item_id.toString())
   }
   return itemIdDict.get(lang).get(name)
 }
@@ -141,14 +161,26 @@ const getItemId = async (lang, name) => {
 // get item id
 const fixUigfJson = (importData) => {
   // if item_id is missing, add placeholder item_id to importData
-  importData.list.forEach(e => {
-    if (!e.item_id) {
-      e.item_id = ""
+  try {
+    if (importData.list) { // v3.0
+      importData.list.forEach(item => {
+        if (!item.item_id) {
+          item.item_id = ""
+        }
+      })
+    } else if (importData.hk4e) { // v4.x
+      importData.hk4e.forEach(account => account.list.forEach((item) => {
+        if (!item.item_id) {
+          item.item_id = ""
+        }
+      }))
     }
-  })
+  } catch (e) {
+    console.error(`Failed to fix UIGF Json: ${e}`)
+  }
 }
 
-const uigfJson = async () => {
+const generateUigf30Json = async () => {
   const { dataMap, current } = getData()
   const data = dataMap.get(current)
   if (!data?.result.size) {
@@ -197,92 +229,158 @@ const uigfJson = async () => {
   return result
 }
 
-const uigf4_2Json = async () => {
+const generateUigf41Json = async (uigfAllAccounts=true) => {
   const { dataMap, current } = getData()
-  const data = dataMap.get(current)
-  if (!data?.result.size) {
+  if (!Array.from(dataMap.entries()).reduce((accumulate, account) => accumulate + account[1].result.size, 0)) {
     throw new Error('数据为空')
   }
-  const fakeId = fakeIdFn()
   const result = {
     info: {
       export_timestamp: Math.round(Date.now() / 1000),
-      export_app: 'genshin-wish-export',
+      export_app: `genshin-wish-export`,
       export_app_version: `v${version}`,
-      version: 'v4.2',
+      version: "v4.1"
     },
-    hk4e: [{
-      uid: data.uid,
-      timezone: data.uid.startsWith('6') ? -5 : data.uid.startsWith('7') ? 1 : 8,
-      lang: data.lang,
-      list: []
-    }],
-    hk4e_ugc: [{
-      uid: data.uid,
-      timezone: data.uid.startsWith('6') ? -5 : data.uid.startsWith('7') ? 1 : 8,
-      lang: data.lang,
-      list: []
-    }]
+    hk4e: []
   }
-  const listHk4eTemp = []
-  const listHk4eUgcTemp = []
-  const uigfLang = uigfLangMap.get(data.lang) || uigfLangMap.get(fixLocalMap.get(data.lang))
-  for (let [type, arr] of data.result) {
-    if (type == '1000' || type == '2000') {
-      for (let item of arr) {
-        listHk4eUgcTemp.push({
-          id: item[5],
-          schedule_id: item[6],
-          op_gacha_type: shouldBeString(item[4]) || type,
-          item_id: item[7],
-          time: item[0],
-          item_name: item[1],
-          item_type: item[2],
-          rank_type: `${item[3]}`,
-        })
+  for (const [uid, data] of dataMap) {
+    if (!uigfAllAccounts && uid != current) continue
+    const uidMatch = uid.match(/^(?<prefix>\d{1,2})\d{8}$/) // match 1 or 2 digits followed by exactly 8 digits
+    const uidPrefix = uidMatch.groups.prefix
+    const uigfLang = uigfLangMap.get(data.lang) || uigfLangMap.get(fixLocalMap.get(data.lang))
+    const fakeId = fakeIdFn()
+    const account = {
+      uid: uid,
+      timezone: ['6'].includes(uidPrefix) ? -5 : ['7'].includes(uidPrefix) ? 1 : 8, // 6(America): TZ=-5, 7(Europe): TZ=1, All others TZ=8
+      lang: uigfRevLangMap.get(data.lang) || data.lang, // ensure long-format for lang
+      list: []
+    }
+    for (const [gachaType, gachaList] of data.result) {
+      for (const gacha of gachaList){
+        const gachaItem = {
+          uigf_gacha_type: gachaType,
+          gacha_type: shouldBeString(gacha[4]) || gachaType,
+          item_id: await getItemId(uigfLang, gacha[1]),
+          // count: null, // optional and not included in stored data
+          time: gacha[0],
+          timestamp: new Date(gacha[0]).getTime(), // Used to sort list for in-order fakeIds
+          name: gacha[1],
+          item_type: gacha[2],
+          rank_type: `${gacha[3]}`,
+          id: shouldBeString(gacha[5]) || ''
+        }
+        account.list.push(gachaItem)
       }
-      continue
     }
-    for (let item of arr) {
-      listHk4eTemp.push({
-        uigf_gacha_type: type,
-        gacha_type: shouldBeString(item[4]) || type,
-        item_id: await getItemId(uigfLang, item[1]),
-        count: "1",
-        time: item[0],
-        name: item[1],
-        item_type: item[2],
-        rank_type: `${item[3]}`,
-        id: shouldBeString(item[5]) || '',
-      })
+    const sortedList = account.list.sort((itemA, itemB) => itemA.timestamp - itemB.timestamp)
+    for (const gachaItem of sortedList) {
+      delete gachaItem.timestamp
+      gachaItem.id = gachaItem.id || fakeId()
     }
+    result.hk4e.push(account)
   }
-  listHk4eTemp.sort((a, b) => a.timestamp - b.timestamp)
-  listHk4eTemp.forEach(item => {
-    delete item.timestamp
-    result.hk4e[0].list.push({
-      ...item,
-      id: item.id || fakeId()
-    })
-  })
-  listHk4eUgcTemp.sort((a, b) => a.timestamp - b.timestamp)
-  listHk4eUgcTemp.forEach(item => {
-    delete item.timestamp
-    result.hk4e_ugc[0].list.push({
-      ...item,
-      id: item.id || fakeId()
-    })
-  })
   return result
 }
 
-const start = async () => {
+const generateUigf42Json = async (uigfAllAccounts = true) => {
+  const { dataMap, current } = getData()
+  if (!Array.from(dataMap.entries()).reduce((accumulate, account) => accumulate + account[1].result.size, 0)) {
+    throw new Error('数据为空')
+  }
+  const result = {
+    info: {
+      export_timestamp: Math.round(Date.now() / 1000),
+      export_app: `genshin-wish-export`,
+      export_app_version: `v${version}`,
+      version: "v4.2"
+    },
+    hk4e: [],
+    hk4e_ugc: []
+  }
+  for (const [uid, data] of dataMap) {
+    if (!uigfAllAccounts && uid != current) continue
+    const uidMatch = uid.match(/^(?<prefix>\d{1,2})\d{8}$/) // match 1 or 2 digits followed by exactly 8 digits
+    const uidPrefix = uidMatch.groups.prefix
+    const uigfLang = uigfLangMap.get(data.lang) || uigfLangMap.get(fixLocalMap.get(data.lang))
+    const fakeId = fakeIdFn()
+    const account = {
+      uid: uid,
+      timezone: ['6'].includes(uidPrefix) ? -5 : ['7'].includes(uidPrefix) ? 1 : 8, // 6(America): TZ=-5, 7(Europe): TZ=1, All others TZ=8
+      lang: uigfRevLangMap.get(data.lang) || data.lang, // ensure long-format for lang
+      list: []
+    }
+    const ugc_account = { ...account, list: [] }
+
+    for (const [gachaType, gachaList] of data.result) {
+      for (const gacha of gachaList) {
+        if (gachaType == '1000' || gachaType == '2000') {
+          const gachaItem = {
+            id: shouldBeString(gacha[5]) || '',
+            op_gacha_type: shouldBeString(gacha[4]) || gachaType,
+            time: gacha[0],
+            item_name: gacha[1],
+            item_type: gacha[2],
+            rank_type: `${gacha[3]}`,
+          }
+          ugc_account.list.push(gachaItem)
+        } else {
+          const gachaItem = {
+            uigf_gacha_type: gachaType,
+            gacha_type: shouldBeString(gacha[4]) || gachaType,
+            item_id: await getItemId(uigfLang, gacha[1]),
+            // count: null, // optional and not included in stored data
+            time: gacha[0],
+            timestamp: new Date(gacha[0]).getTime(), // Used to sort list for in-order fakeIds
+            name: gacha[1],
+            item_type: gacha[2],
+            rank_type: `${gacha[3]}`,
+            id: shouldBeString(gacha[5]) || ''
+          }
+          account.list.push(gachaItem)
+        }
+      }
+    }
+    const hk4eSortedList = account.list.sort((itemA, itemB) => itemA.timestamp - itemB.timestamp)
+    for (const gachaItem of hk4eSortedList) {
+      delete gachaItem.timestamp
+      gachaItem.id = gachaItem.id || fakeId()
+    }
+    result.hk4e.push(account)
+    const hk4eUgcSortedList = ugc_account.list.sort((itemA, itemB) => itemA.timestamp - itemB.timestamp)
+    for (const gachaItem of hk4eUgcSortedList) {
+      delete gachaItem.timestamp
+      gachaItem.id = gachaItem.id || fakeId()
+    }
+    result.hk4e_ugc.push(ugc_account)
+  }
+  return result
+}
+
+const start = async (uigfVersion, uigfAllAccounts=true) => {
   await initLookupTable()
-  // const result = await uigfJson()
-  const result = await uigf4_2Json()
+  // const result = uigfVersion === "3.0" ? await generateUigf30Json() : await generateUigf41Json(uigfAllAccounts)
+  // const uid = uigfVersion === '3.0' ? result.info.uid : result.hk4e[0].uid
+  // const numAccounts = uigfVersion === '3.0' ? 1 : result.hk4e.length
+  let result
+  let uid
+  let numAccounts
+  if (uigfVersion === "3.0") {
+    result = await generateUigf30Json()
+    uid = result.info.uid
+    numAccounts = 1
+  } else if (uigfVersion === "4.1") {
+    result = await generateUigf41Json(uigfAllAccounts)
+    uid = result.hk4e[0].uid
+    numAccounts = result.hk4e.length
+  } else {
+    result = await generateUigf42Json(uigfAllAccounts)
+    uid = result.hk4e[0].uid
+    numAccounts = result.hk4e.length
+  }
   await saveLookupTable()
+  const uigfFileName = `UIGF_v${uigfVersion}` + (numAccounts > 1 ? '' : `_${uid}`) + `_${getTimeString()}.json`
   const filePath = dialog.showSaveDialogSync({
-    defaultPath: path.join(app.getPath('downloads'), `UIGF_${result.hk4e[0].uid}_${getTimeString()}`),
+    defaultPath: path.join(app.getPath('downloads'), uigfFileName),
     filters: [
       { name: 'JSON', extensions: ['json'] }
     ]
@@ -301,6 +399,105 @@ const saveAndBackup = async (data) => {
   }
   await saveData(data)
   await changeCurrent(data.uid)
+}
+
+const importUgif30Json = async (importData) => {
+    const gachaData = {
+      result: new Map(),
+      time: Date.now(),
+      typeMap: getItemTypeNameMap(importData.info.lang),
+      uid: importData.info.uid,
+      lang: importData.info.lang
+    }
+    gachaData.typeMap.forEach((_, k) => gachaData.result.set(k, []))
+    importData.list.sort((a, b) => parseInt(BigInt(a.id) - BigInt(b.id)))
+    for (const item of importData.list) {
+      gachaData.result.get(item.uigf_gacha_type).push([
+        item.time,
+        item.name,
+        item.item_type,
+        parseInt(item.rank_type),
+        item.gacha_type,
+        item.id 
+      ])
+    }
+    await saveAndBackup(gachaData)
+}
+
+const importUgif41Json = async (importData) => {
+  for (const accountData of importData.hk4e) {
+      const gachaData = {
+        result: new Map(),
+        time: Date.now(),
+        typeMap: getItemTypeNameMap(accountData.lang),
+        uid: accountData.uid,
+        lang: accountData.lang
+      }
+      gachaData.typeMap.forEach((_, k) => gachaData.result.set(k, []))
+      accountData.list.sort((itemA, itemB) => parseInt(BigInt(itemA.id) - BigInt(itemB.id)))
+      for (const item of accountData.list) {
+        const gachaItem = [
+          item.time,
+          item.name,
+          item.item_type,
+          parseInt(item.rank_type),
+          item.gacha_type,
+          item.id
+        ]
+        gachaData.result.get(item.uigf_gacha_type).push(gachaItem)
+      }
+      await saveAndBackup(gachaData)
+  }
+}
+
+const importUgif42Json = async (importData) => {
+  const accountsMap = new Map()
+
+  for (const accountData of importData.hk4e) {
+      const gachaData = {
+        result: new Map(),
+        time: Date.now(),
+        typeMap: getItemTypeNameMap(accountData.lang),
+        uid: accountData.uid,
+        lang: accountData.lang
+      }
+      gachaData.typeMap.forEach((_, k) => gachaData.result.set(k, []))
+      accountData.list.sort((itemA, itemB) => parseInt(BigInt(itemA.id) - BigInt(itemB.id)))
+      for (const item of accountData.list) {
+        const gachaItem = [
+          item.time,
+          item.name,
+          item.item_type,
+          parseInt(item.rank_type),
+          item.gacha_type,
+          item.id
+        ]
+        gachaData.result.get(item.uigf_gacha_type).push(gachaItem)
+      }
+      accountsMap.set(accountData.uid, gachaData)
+  }
+  for (const accountData of importData.hk4e_ugc) {
+      const gachaData = accountsMap.get(accountData.uid)
+      if (!gachaData) continue
+      for (const item of accountData.list) {
+        const gachaItem = [
+          item.time,
+          item.item_name,
+          item.item_type,
+          parseInt(item.rank_type),
+          item.op_gacha_type,
+          item.id
+        ]
+        if (item.op_gacha_type == '1000'){
+          gachaData.result.get(item.op_gacha_type).push(gachaItem)
+        } else {
+          gachaData.result.get('2000').push(gachaItem)
+        }
+      }
+      for (const gachaData of accountsMap.values()) {
+        await saveAndBackup(gachaData)
+      }
+  }
 }
 
 const importJson = async () => {
@@ -323,20 +520,15 @@ const importJson = async () => {
       // try to fix imported data when possible
       fixUigfJson(importData)
       // then validate imported data using schema
-      if (validateUigfJson(importData)) {
-        const gachaData = {
-          result: new Map(),
-          time: Date.now(),
-          typeMap: getItemTypeNameMap(importData.info.lang),
-          uid: importData.info.uid,
-          lang: importData.info.lang
+      if (validateUigf30Json(importData)) {
+        await importUgif30Json(importData)
+      } else if (validateUigf41Json(importData)) {
+        const uigfVersion = importData.info?.version
+        if (uigfVersion === 'v4.2') {
+          await importUgif42Json(importData)
+          return
         }
-        gachaData.typeMap.forEach((_, k) => gachaData.result.set(k, []))
-        importData.list.sort((a, b) => parseInt(BigInt(a.id) - BigInt(b.id)))
-        for (const item of importData.list) {
-          gachaData.result.get(item.uigf_gacha_type).push([item.time, item.name, item.item_type, parseInt(item.rank_type), item.gacha_type, item.id])
-        }
-        await saveAndBackup(gachaData)
+        await importUgif41Json(importData)
       } else {
         throw new Error(`JSON format error`)
       }
@@ -346,12 +538,12 @@ const importJson = async () => {
   }
 }
 
-ipcMain.handle('EXPORT_UIGF_JSON', async () => {
-  await start()
+ipcMain.handle('EXPORT_UIGF_JSON', async (event, uigfVersion, uigfAllAccounts=true) => {
+  await start(uigfVersion, uigfAllAccounts)
 })
 
 ipcMain.handle('IMPORT_UIGF_JSON', async () => {
   return await importJson()
 })
 
-module.exports = { uigfJson, uigf4_2Json }
+module.exports = { generateUigf30Json, generateUigf41Json, generateUigf42Json }